+++ src/main/resources/static/css/style.css
@@ -234,3 +234,8 @@ li.movie-search {
 h1.test {
     font-weight: bolder !important;
 }
+
+/* 영화 상세페이지 */
+#movieInfo .infoItem {
+    margin: 10px 10px;
+}
